i have a problem of printing after a death of the first philosofer

+++ add_philosofers.c
@@ -6,7 +6,7 @@
 /*   By: mskhairi <mskhairi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/21 20:34:58 by mskhairi          #+#    #+#             */
-/*   Updated: 2024/06/21 20:36:10 by mskhairi         ###   ########.fr       */
+/*   Updated: 2024/06/22 12:33:08 by mskhairi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,13 @@
 
 void	add_philosofers(t_data *data, t_philo *philo)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	data->mutex = malloc(data->philos_nbr * sizeof(pthread_mutex_t));
 	while (i < data->philos_nbr)
 	{
+        // philo[i].last_time2eat = data->start_time;
 		if (pthread_mutex_init(&data->mutex[i], NULL) != 0)
 			printf("error initing mutex\n");
 		philo[i].data = data;
